Adjust background of linear picker views
Problem:
The backgrounds of the brightness and saturation picker views should show the correct colors if the other slider has been moved

Analysis:
The linear picker views have a mask that are used as overlay to display the gradient and the background colors are set to display the hue value only

Solution:
Instead of using a layer mask, the frontLayer is now not transparent anymore but is using the current background color as gradient endpoint. Whenever the background color changes this gradient is recalculated to match the selected colors. In the CircleColorPickerView, the background colors are now set to match the opposite (saturation -> fixed brightness, brightness ->  fixed saturation). The bubble views use the current color so get a nice finish.